requirements.txt
absl-py==0.8.0
astor==0.8.0
attrdict==2.0.1
attrs==19.3.0
backcall==0.1.0
biwrap==0.1.6
bleach==3.1.0
cachetools==3.1.1
carla==0.9.8
certifi==2019.11.28
chardet==3.0.4
cloudpickle==1.1.1
cycler==0.10.0
dataclasses==0.7
DateTime==4.3
decorator==4.4.0
defusedxml==0.6.0
descartes==1.1.0
dill==0.3.1.1
entrypoints==0.3
fasteners==0.15
fire==0.3.1
gast==0.2.2
gin-config==0.4.0
gitdb2==2.0.6
GitPython==3.0.3
google-api-python-client==1.7.11
google-auth==1.7.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
grpcio==1.24.1
h5py==2.10.0
httplib2==0.14.0
Hydra==2.5
hydra-core==0.11.2
idna==2.8
imageio==2.6.1
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.0
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
monotonic==1.5
more-itertools==7.2.0
motmetrics==1.1.3
natsort==7.1.1
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.4
notebook==6.0.1
numpy==1.17.2
nuscenes-devkit==1.0.8
oauthlib==3.1.0
omegaconf==1.4.0
opencv-python==4.1.1.26
opt-einsum==3.1.0
pandas==1.0.3
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
pptk==0.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.10.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pygame==1.9.6
Pygments==2.4.2
pygsheets==2.0.2
pyparsing==2.4.2
pyquaternion==0.9.5
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.3.2
screeninfo==0.6.1
Send2Trash==1.5.0
Shapely==1.7.0
six==1.12.0
sklearn==0.0
smmap2==2.0.5
tensorboard==1.14.0
tensorflow-addons==0.8.3
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
tensorflow-plot==0.3.2
tensorflow-probability==0.7.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
torch==1.4.0
torchvision==0.5.0
tornado==6.0.3
tqdm==4.38.0
traitlets==4.3.3
typeguard==2.7.1
uritemplate==3.0.0
urllib3==1.25.7
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
zope.interface==5.1.0
